Extract ChannelAllocator
Extracts the logic of a thread which allocates channels on
the bridge and invites participant from JitsiMeetConference
to ChannelAllocator class.
