+++ src/main/java/org/jitsi/jicofo/JitsiMeetConference.java
@@ -19,8 +19,6 @@ package org.jitsi.jicofo;
 
 import net.java.sip.communicator.impl.protocol.jabber.extensions.colibri.*;
 import net.java.sip.communicator.impl.protocol.jabber.extensions.jingle.*;
-import net.java.sip.communicator.impl.protocol.jabber.extensions.jitsimeet.*;
-import net.java.sip.communicator.impl.protocol.jabber.jinglesdp.*;
 import net.java.sip.communicator.service.protocol.*;
 import net.java.sip.communicator.service.protocol.event.*;
 import net.java.sip.communicator.util.*;
@@ -28,15 +26,12 @@ import net.java.sip.communicator.util.Logger;
 
 import org.jitsi.assertions.*;
 import org.jitsi.impl.protocol.xmpp.extensions.*;
-import org.jitsi.jicofo.discovery.*;
 import org.jitsi.jicofo.event.*;
 import org.jitsi.jicofo.log.*;
 import org.jitsi.jicofo.reservation.*;
-import org.jitsi.jicofo.util.*;
 import org.jitsi.protocol.xmpp.*;
 import org.jitsi.protocol.xmpp.colibri.*;
 import org.jitsi.protocol.xmpp.util.*;
-import org.jitsi.service.neomedia.*;
 import org.jitsi.util.*;
 import org.jitsi.eventadmin.*;
 
@@ -67,13 +62,6 @@ public class JitsiMeetConference
      */
     private final static Logger logger
         = Logger.getLogger(JitsiMeetConference.class);
-
-    /**
-     * Error code used in {@link OperationFailedException} when there are no
-     * working videobridge bridges.
-     * FIXME: consider moving to OperationFailedException ?
-     */
-    private final static int BRIDGE_FAILURE_ERR_CODE = 20;
 
     /**
      * Format used to print the date into the focus identifier string.
@@ -123,18 +111,6 @@ public class JitsiMeetConference
      * flag to skip channel expiration step when conference is being disposed.
      */
     private boolean bridgeHasFailed;
-
-    /**
-     * Synchronization root for currently selected bridge for the
-     * {@link #colibriConference}.
-     */
-    private final Object bridgeSelectSync = new Object();
-
-    /**
-     * The instance of <tt>BridgeSelector</tt> we use to select JVB for this
-     * conference.
-     */
-    private BridgeSelector bridgeSelector;
 
     /**
      * The name of XMPP user used by the focus to login.
@@ -241,6 +217,7 @@ public class JitsiMeetConference
                                JitsiMeetGlobalConfig    globalConfig)
     {
         Assert.notNull(protocolProviderHandler, "protocolProviderHandler");
+        Assert.notNull(config, "config");
 
         this.id = ID_DATE_FORMAT.format(new Date()) + "_" + hashCode();
         this.roomName = roomName;
@@ -291,15 +268,13 @@ public class JitsiMeetConference
 
             recording = new JitsiMeetRecording(this, services);
 
-            bridgeSelector = services.getBridgeSelector();
-
             // Set pre-configured videobridge
             String preConfiguredBridge = config.getPreConfiguredVideobridge();
 
             if (!StringUtils.isNullOrEmpty(preConfiguredBridge))
             {
-                bridgeSelector.setPreConfiguredBridge(
-                        preConfiguredBridge);
+                services.getBridgeSelector()
+                    .setPreConfiguredBridge(preConfiguredBridge);
             }
 
             // Set pre-configured SIP gateway
@@ -468,14 +443,14 @@ public class JitsiMeetConference
                     = hasToStartMuted(
                             member, member == chatRoomMember /* justJoined */);
 
-                inviteChatMember(member, startMuted);
+                inviteChatMember(member, startMuted, colibriConference);
             }
         }
         // Only the one who has just joined
         else
         {
             final boolean[] startMuted = hasToStartMuted(chatRoomMember, true);
-            inviteChatMember(chatRoomMember, startMuted);
+            inviteChatMember(chatRoomMember, startMuted, colibriConference);
         }
     }
 
@@ -487,8 +462,9 @@ public class JitsiMeetConference
      * @param startMuted array with values for audio and video that indicates
      * whether the participant should start muted.
      */
-    private void inviteChatMember(final ChatRoomMember    chatRoomMember,
-                                  final boolean[]         startMuted)
+    private void inviteChatMember(final ChatRoomMember       chatRoomMember,
+                                  final boolean[]            startMuted,
+                                  final ColibriConference    colibriConference)
     {
         if (isFocusMember(chatRoomMember))
             return;
@@ -512,22 +488,9 @@ public class JitsiMeetConference
 
         // Invite peer takes time because of channel allocation, so schedule
         // this on separate thread.
-        FocusBundleActivator.getSharedThreadPool().submit(new Runnable()
-        {
-            @Override
-            public void run()
-            {
-                try
-                {
-                    discoverFeaturesAndInvite(
-                            newParticipant, address, startMuted);
-                }
-                catch (Exception e)
-                {
-                    logger.error("Exception on participant invite", e);
-                }
-            }
-        });
+        FocusBundleActivator.getSharedThreadPool().submit(
+                new ChannelAllocator(
+                        this, colibriConference, newParticipant, startMuted));
     }
 
     /**
@@ -584,511 +547,6 @@ public class JitsiMeetConference
         }
 
         return startMuted;
-    }
-
-    /**
-     * Methods executed on thread pool is time consuming as is doing feature
-     * discovery and channel allocation for new participants.
-     * @param newParticipant new <tt>Participant</tt> instance.
-     * @param address new participant full MUC address.
-     * @param startMuted if the first element is <tt>true</tt> the participant
-     * will start audio muted. if the second element is <tt>true</tt> the
-     * participant will start video muted.
-     */
-    private void discoverFeaturesAndInvite(Participant     newParticipant,
-                                           String          address,
-                                           boolean[]       startMuted)
-    {
-        // Feature discovery
-        List<String> features
-            = DiscoveryUtil.discoverParticipantFeatures(
-                    getXmppProvider(), address);
-
-        newParticipant.setSupportedFeatures(features);
-
-        logger.info(
-            address + " has bundle ? " + newParticipant.hasBundleSupport());
-
-        // Store instance here as it is set to null when conference is disposed
-        ColibriConference conference = this.colibriConference;
-        List<ContentPacketExtension> offer;
-
-        try
-        {
-            offer = createOffer(newParticipant);
-            if (offer == null)
-            {
-                logger.info("Channel allocation cancelled for " + address);
-                return;
-            }
-        }
-        catch (OperationFailedException e)
-        {
-            //FIXME: retry ? sometimes it's just timeout
-            logger.error("Failed to allocate channels for " + address, e);
-
-            // Notify users about bridge is down event
-            if (BRIDGE_FAILURE_ERR_CODE == e.getErrorCode() && chatRoom != null)
-            {
-                meetTools.sendPresenceExtension(
-                    chatRoom, new BridgeIsDownPacketExt());
-            }
-            // Cancel - no channels allocated
-            return;
-        }
-        /*
-           This check makes sure that at the point when we're trying to
-           invite new participant:
-           - the conference has not been disposed in the meantime
-           - he's still in the room
-           - we have managed to send Jingle session-initiate
-           Otherwise we expire allocated channels.
-        */
-        boolean expireChannels = false;
-
-        if (chatRoom == null)
-        {
-            // Conference disposed
-            logger.info(
-                "Expiring " + address + " channels - conference disposed");
-
-            expireChannels = true;
-        }
-        else if (findMember(address) == null)
-        {
-            // Participant has left the room
-            logger.info(
-                "Expiring " + address + " channels - participant has left");
-
-            expireChannels = true;
-        }
-        else if (!jingle.initiateSession(
-                     newParticipant.hasBundleSupport(), address, offer, this,
-                     startMuted))
-        {
-            // Failed to invite
-            logger.info(
-                "Expiring " + address +
-                    " channels - no RESULT for session-invite");
-
-            expireChannels = true;
-        }
-
-        if (expireChannels)
-        {
-            conference.expireChannels(
-                newParticipant.getColibriChannelsInfo());
-        }
-    }
-
-    /**
-     * Allocates Colibri channels for given {@link Participant} by trying all
-     * available bridges returned by {@link BridgeSelector}.
-     *
-     * @param peer the for whom Colibri channel are to be allocated.
-     * @param contents the media offer description passed to the bridge.
-     *
-     * @return {@link ColibriConferenceIQ} that describes channels allocated for
-     *         given <tt>peer</tt>. <tt>null</tt> is returned if conference is
-     *         disposed before we manage to allocate the channels.
-     *
-     * @throws OperationFailedException if we have failed to allocate channels
-     *         using existing bridge and we can not switch to another bridge.
-     */
-    private ColibriConferenceIQ allocateChannels(
-            Participant                     peer,
-            List<ContentPacketExtension>    contents)
-        throws OperationFailedException
-    {
-        // This method is executed on thread pool.
-
-        // Store colibri instance here to be able to free the channels even
-        // after the conference has been disposed.
-        ColibriConference colibriConference = this.colibriConference;
-        if (this.colibriConference == null)
-        {
-            // Nope - the conference has been disposed, before the thread got
-            // the chance to do anything
-            return null;
-        }
-
-        // Set initial bridge if we haven't used any yet
-        synchronized (bridgeSelectSync)
-        {
-            if (StringUtils.isNullOrEmpty(
-                    colibriConference.getJitsiVideobridge()))
-            {
-                String bridge = null;
-
-                if (!StringUtils.isNullOrEmpty(config.getEnforcedVideobridge()))
-                {
-                    bridge = config.getEnforcedVideobridge();
-                    logger.info(
-                        "Will force bridge: " + bridge
-                            + " on: " + getRoomName());
-                }
-                else
-                {
-                    bridge = bridgeSelector.selectVideobridge();
-                }
-
-                if (StringUtils.isNullOrEmpty(bridge))
-                {
-                    throw new OperationFailedException(
-                        "Failed to allocate channels - no bridge configured",
-                        BRIDGE_FAILURE_ERR_CODE);
-                }
-
-                colibriConference.setJitsiVideobridge(bridge);
-            }
-        }
-
-        String jvb = null;
-
-        while (this.colibriConference != null)
-        {
-            try
-            {
-                synchronized (bridgeSelectSync)
-                {
-                    jvb = colibriConference.getJitsiVideobridge();
-                }
-
-                logger.info(
-                    "Using " + jvb + " to allocate channels for: "
-                        + peer.getChatMember().getContactAddress());
-
-                ColibriConferenceIQ peerChannels
-                    = colibriConference.createColibriChannels(
-                            peer.hasBundleSupport(),
-                            peer.getEndpointId(),
-                            true, contents);
-
-                bridgeSelector.updateBridgeOperationalStatus(jvb, true);
-
-                if (colibriConference.hasJustAllocated())
-                {
-                    EventAdmin eventAdmin
-                            = FocusBundleActivator.getEventAdmin();
-                    if (eventAdmin != null)
-                    {
-                        eventAdmin.sendEvent(
-                            EventFactory.conferenceRoom(
-                                    colibriConference.getConferenceId(),
-                                    roomName,
-                                    getId(),
-                                    jvb));
-                    }
-
-                    if (recording != null)
-                        recording.onConferenceAllocated();
-                }
-                return peerChannels;
-            }
-            catch(OperationFailedException exc)
-            {
-                logger.error(
-                    "Failed to allocate channels using bridge: " + jvb, exc);
-
-                bridgeSelector.updateBridgeOperationalStatus(jvb, false);
-
-                // Check if the conference is in progress
-                if (!StringUtils.isNullOrEmpty(
-                        colibriConference.getConferenceId()))
-                {
-                    // Restart
-                    logger.error("Bridge failure - stopping the conference");
-                    stop();
-                }
-
-                // Try next bridge
-                synchronized (bridgeSelectSync)
-                {
-                    if (StringUtils.isNullOrEmpty(
-                            config.getEnforcedVideobridge()))
-                    {
-                        jvb = bridgeSelector.selectVideobridge();
-                    }
-                    else
-                    {
-                        // If the "enforced" bridge has failed we do not try
-                        // any other bridges, but fail immediately
-                        jvb = null;
-                    }
-
-                    // Is it the same which has just failed ?
-                    // (we do not always call iterator.next() at the beginning)
-                    //if (faultyBridge.equals(nextBridge))
-                    //    nextBridge = null;
-                    if (!StringUtils.isNullOrEmpty(jvb))
-                    {
-                        colibriConference.setJitsiVideobridge(jvb);
-                    }
-                    else
-                    {
-                        // No more bridges to try
-                        throw new OperationFailedException(
-                            "Failed to allocate channels " +
-                                "- all bridges are faulty",
-                                BRIDGE_FAILURE_ERR_CODE);
-                    }
-                }
-            }
-        }
-        // If we reach this point it means that the conference has been disposed
-        // before we've managed to allocate anything
-        return null;
-    }
-
-    /**
-     * Creates Jingle offer for given {@link Participant}.
-     *
-     * @param peer the participant for whom Jingle offer will be created.
-     *
-     * @return the list of contents describing conference Jingle offer or
-     *         <tt>null</tt> if the conference has been disposed before we've
-     *         managed to allocate Colibri channels.
-     *
-     * @throws OperationFailedException if focus fails to allocate channels
-     *         or something goes wrong.
-     */
-    private List<ContentPacketExtension> createOffer(Participant peer)
-        throws OperationFailedException
-    {
-        List<ContentPacketExtension> contents = new ArrayList<>();
-
-        boolean disableIce = !peer.hasIceSupport();
-        boolean useDtls = peer.hasDtlsSupport();
-
-        if (peer.hasAudioSupport())
-        {
-            contents.add(
-                JingleOfferFactory.createContentForMedia(
-                    MediaType.AUDIO, disableIce, useDtls));
-        }
-
-        if (peer.hasVideoSupport())
-        {
-            contents.add(
-                JingleOfferFactory.createContentForMedia(
-                    MediaType.VIDEO, disableIce, useDtls));
-        }
-
-        // Is SCTP enabled ?
-        boolean openSctp = config == null || config.openSctp() == null
-                ? true : config.openSctp();
-
-        if (openSctp && peer.hasSctpSupport())
-        {
-            contents.add(
-                JingleOfferFactory.createContentForMedia(
-                    MediaType.DATA, disableIce, useDtls));
-        }
-
-        boolean useBundle = peer.hasBundleSupport();
-
-        ColibriConferenceIQ peerChannels = allocateChannels(peer, contents);
-
-        if (peerChannels == null)
-            return null;
-
-        peer.setColibriChannelsInfo(peerChannels);
-
-        for (ContentPacketExtension cpe : contents)
-        {
-            ColibriConferenceIQ.Content colibriContent
-                = peerChannels.getContent(cpe.getName());
-
-            if (colibriContent == null)
-                continue;
-
-            // Channels
-            for (ColibriConferenceIQ.Channel channel
-                : colibriContent.getChannels())
-            {
-                IceUdpTransportPacketExtension transport;
-
-                if (useBundle)
-                {
-                    ColibriConferenceIQ.ChannelBundle bundle
-                        = peerChannels.getChannelBundle(
-                                channel.getChannelBundleId());
-
-                    if (bundle == null)
-                    {
-                        logger.error(
-                            "No bundle for " + channel.getChannelBundleId());
-                        continue;
-                    }
-
-                    transport = bundle.getTransport();
-
-                    if (!transport.isRtcpMux())
-                    {
-                        transport.addChildExtension(
-                            new RtcpmuxPacketExtension());
-                    }
-                }
-                else
-                {
-                    transport = channel.getTransport();
-                }
-
-                try
-                {
-                    // Remove empty transport
-                    IceUdpTransportPacketExtension empty
-                        = cpe.getFirstChildOfType(
-                                IceUdpTransportPacketExtension.class);
-                    cpe.getChildExtensions().remove(empty);
-
-                    cpe.addChildExtension(
-                        IceUdpTransportPacketExtension
-                            .cloneTransportAndCandidates(transport, true));
-                }
-                catch (Exception e)
-                {
-                    logger.error(e, e);
-                }
-            }
-            // SCTP connections
-            for (ColibriConferenceIQ.SctpConnection sctpConn
-                : colibriContent.getSctpConnections())
-            {
-                IceUdpTransportPacketExtension transport;
-
-                if (useBundle)
-                {
-                    ColibriConferenceIQ.ChannelBundle bundle
-                        = peerChannels.getChannelBundle(
-                                sctpConn.getChannelBundleId());
-
-                    if (bundle == null)
-                    {
-                        logger.error(
-                            "No bundle for " + sctpConn.getChannelBundleId());
-                        continue;
-                    }
-
-                    transport = bundle.getTransport();
-                }
-                else
-                {
-                    transport = sctpConn.getTransport();
-                }
-
-                try
-                {
-                    // Remove empty transport
-                    IceUdpTransportPacketExtension empty
-                        = cpe.getFirstChildOfType(
-                                IceUdpTransportPacketExtension.class);
-                    cpe.getChildExtensions().remove(empty);
-
-                    IceUdpTransportPacketExtension copy
-                        = IceUdpTransportPacketExtension
-                            .cloneTransportAndCandidates(transport, true);
-
-                    // FIXME: hardcoded
-                    SctpMapExtension sctpMap = new SctpMapExtension();
-                    sctpMap.setPort(5000);
-                    sctpMap.setProtocol(
-                            SctpMapExtension.Protocol.WEBRTC_CHANNEL);
-                    sctpMap.setStreams(1024);
-
-                    copy.addChildExtension(sctpMap);
-
-                    cpe.addChildExtension(copy);
-                }
-                catch (Exception e)
-                {
-                    logger.error(e, e);
-                }
-            }
-            // Existing peers SSRCs
-            RtpDescriptionPacketExtension rtpDescPe
-                = JingleUtils.getRtpDescription(cpe);
-            if (rtpDescPe != null)
-            {
-                if (useBundle)
-                {
-                    // rtcp-mux
-                    rtpDescPe.addChildExtension(
-                        new RtcpmuxPacketExtension());
-                }
-
-                // Copy SSRC sent from the bridge(only the first one)
-                for (ColibriConferenceIQ.Channel channel
-                    : colibriContent.getChannels())
-                {
-                    SourcePacketExtension ssrcPe
-                        = channel.getSources().size() > 0
-                                ? channel.getSources().get(0) : null;
-                    if (ssrcPe == null)
-                        continue;
-
-                    try
-                    {
-                        String contentName = colibriContent.getName();
-                        SourcePacketExtension ssrcCopy = ssrcPe.copy();
-
-                        // FIXME: not all parameters are used currently
-                        ssrcCopy.addParameter(
-                            new ParameterPacketExtension("cname","mixed"));
-                        ssrcCopy.addParameter(
-                            new ParameterPacketExtension(
-                                "label",
-                                "mixedlabel" + contentName + "0"));
-                        ssrcCopy.addParameter(
-                            new ParameterPacketExtension(
-                                "msid",
-                                "mixedmslabel mixedlabel"
-                                    + contentName + "0"));
-                        ssrcCopy.addParameter(
-                            new ParameterPacketExtension(
-                                "mslabel", "mixedmslabel"));
-
-                        // Mark 'jvb' as SSRC owner
-                        SSRCInfoPacketExtension ssrcInfo
-                            = new SSRCInfoPacketExtension();
-                        ssrcInfo.setOwner("jvb");
-                        ssrcCopy.addChildExtension(ssrcInfo);
-
-                        rtpDescPe.addChildExtension(ssrcCopy);
-                    }
-                    catch (Exception e)
-                    {
-                        logger.error("Copy SSRC error", e);
-                    }
-                }
-
-                // Include all peers SSRCs
-                List<SourcePacketExtension> mediaSources
-                    = getAllSSRCs(cpe.getName());
-
-                for (SourcePacketExtension ssrc : mediaSources)
-                {
-                    try
-                    {
-                        rtpDescPe.addChildExtension(ssrc.copy());
-                    }
-                    catch (Exception e)
-                    {
-                        logger.error("Copy SSRC error", e);
-                    }
-                }
-
-                // Include SSRC groups
-                List<SourceGroupPacketExtension> sourceGroups
-                    = getAllSSRCGroups(cpe.getName());
-                for(SourceGroupPacketExtension ssrcGroup : sourceGroups)
-                {
-                    rtpDescPe.addChildExtension(ssrcGroup);
-                }
-            }
-        }
-
-        return contents;
     }
 
     /**
@@ -1238,14 +696,7 @@ public class JitsiMeetConference
                         leftPeer.getSSRCGroupsCopy(),
                         false /* no JVB update - will expire */);
 
-                ColibriConferenceIQ peerChannels
-                        = leftPeer.getColibriChannelsInfo();
-                if (!bridgeHasFailed && peerChannels != null)
-                {
-                    logger.info("Expiring channels for: " + contactAddress);
-                    colibriConference.expireChannels(
-                        leftPeer.getColibriChannelsInfo());
-                }
+                expireParticipantChannels(colibriConference, leftPeer);
             }
             boolean removed = participants.remove(leftPeer);
             logger.info(
@@ -1259,6 +710,29 @@ public class JitsiMeetConference
         if (participants.size() == 0)
         {
             stop();
+        }
+    }
+
+    /**
+     * Expires channels for given participant unless there are some
+     * circumstances that prevents us from doing it.
+     *
+     * @param colibriConference <tt>ColibriConference</tt> instance that owns
+     *        the channels to be expired.
+     * @param participant the <tt>Participant</tt> for whom Colibri channels are
+     *        to be expired.
+     */
+    void expireParticipantChannels(ColibriConference colibriConference,
+                                   Participant       participant)
+    {
+        ColibriConferenceIQ channelsInfo
+            = participant.getColibriChannelsInfo();
+
+        if (channelsInfo != null && colibriConference != null
+                && !bridgeHasFailed)
+        {
+            logger.info("Expiring channels for: " + participant.getMucJid());
+            colibriConference.expireChannels(channelsInfo);
         }
     }
 
@@ -1579,10 +1053,10 @@ public class JitsiMeetConference
     {
         Participant sourcePeer
             = findParticipantForJingleSession(sourceJingleSession);
+        String peerAddress = sourceJingleSession.getAddress();
         if (sourcePeer == null)
         {
-            logger.error("Remove-source: no session for "
-                             + sourceJingleSession.getAddress());
+            logger.error("Remove-source: no session for " + peerAddress);
             return;
         }
 
@@ -1595,8 +1069,7 @@ public class JitsiMeetConference
         if (removedSSRCs.isEmpty() && removedGroups.isEmpty())
         {
             logger.warn(
-                "No ssrcs or groups to be removed from: "
-                    + sourceJingleSession.getAddress());
+                    "No ssrcs or groups to be removed from: "+ peerAddress);
             return;
         }
 
@@ -1605,7 +1078,11 @@ public class JitsiMeetConference
         ssrcGroupsToRemove = removedGroups;
 
         // Updates SSRC Groups on the bridge
-        if (updateChannels)
+        ColibriConference colibriConference = this.colibriConference;
+        // We may hit null here during conference restart, but that's not
+        // important since the bridge for this instance will not be used
+        // anymore and state is synced up soon after channels are allocated
+        if (updateChannels && colibriConference != null)
         {
             colibriConference.updateSourcesInfo(
                     sourcePeer.getSSRCsCopy(),
@@ -1613,9 +1090,7 @@ public class JitsiMeetConference
                     sourcePeer.getColibriChannelsInfo());
         }
 
-        logger.info(
-            "Removing " + sourceJingleSession.getAddress()
-                + " SSRCs " + ssrcsToRemove);
+        logger.info("Removing " + peerAddress + " SSRCs " + ssrcsToRemove);
 
         for (Participant peer : participants)
         {
@@ -1626,8 +1101,7 @@ public class JitsiMeetConference
             if (jingleSessionToNotify == null)
             {
                 logger.warn(
-                    "Remove source: no jingle session for "
-                        + peer.getChatMember().getContactAddress());
+                        "Remove source: no jingle session for " + peerAddress);
 
                 peer.scheduleSSRCsToRemove(ssrcsToRemove);
 
@@ -1650,7 +1124,7 @@ public class JitsiMeetConference
      * @return the list of all SSRCs of given media type that exist in current
      *         conference state.
      */
-    private List<SourcePacketExtension> getAllSSRCs(String media)
+    List<SourcePacketExtension> getAllSSRCs(String media)
     {
         List<SourcePacketExtension> mediaSSRCs = new ArrayList<>();
 
@@ -1675,7 +1149,7 @@ public class JitsiMeetConference
      * @return the list of all SSRC groups of given media type that exist in
      *         current conference state.
      */
-    private List<SourceGroupPacketExtension> getAllSSRCGroups(String media)
+    List<SourceGroupPacketExtension> getAllSSRCGroups(String media)
     {
         List<SourceGroupPacketExtension> ssrcGroups = new ArrayList<>();
 
@@ -1857,7 +1331,7 @@ public class JitsiMeetConference
      * Handles on bridge down event by shutting down the conference if it's the
      * one we're using here.
      */
-    private void onBridgeDown(String bridgeJid)
+    void onBridgeDown(String bridgeJid)
     {
         if (colibriConference != null
                 && bridgeJid.equals(colibriConference.getJitsiVideobridge()))
@@ -1880,6 +1354,15 @@ public class JitsiMeetConference
     }
 
     /**
+     * Returns config for this conference.
+     * @return <tt>JitsiMeetConfig</tt> instance used in this conference.
+     */
+    public JitsiMeetConfig getConfig()
+    {
+        return config;
+    }
+
+    /**
      * Returns <tt>JitsiMeetRecording</tt> for this conference.
      * @return <tt>JitsiMeetRecording</tt> instance for this conference or
      *         <tt>null</tt> if it's not available yet(should be after we join
@@ -1888,6 +1371,34 @@ public class JitsiMeetConference
     public JitsiMeetRecording getRecording()
     {
         return recording;
+    }
+
+    /**
+     * Methods called by {@link ChannelAllocator} just after it has created new
+     * Colibri conference on the JVB.
+     *
+     * @param colibriConference <tt>ColibriConference</tt> instance which just
+     * has been allocated on the bridge
+     * @param videobridgeJid the JID of the JVB where given
+     * <tt>ColibriConference</tt> has been allocated
+     */
+    public void onColibriConferenceAllocated(
+            ColibriConference    colibriConference,
+            String               videobridgeJid)
+    {
+        EventAdmin eventAdmin = FocusBundleActivator.getEventAdmin();
+        if (eventAdmin != null)
+        {
+            eventAdmin.sendEvent(
+                    EventFactory.conferenceRoom(
+                            colibriConference.getConferenceId(),
+                            roomName,
+                            getId(),
+                            videobridgeJid));
+        }
+
+        if (recording != null)
+            recording.onConferenceAllocated();
     }
 
     /**
